% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsr.R
\name{adsr_filter}
\alias{adsr_filter}
\alias{adsr_filter.wave}
\title{ADSR}
\usage{
adsr_filter(x, attack, decay, sustain, hold, release, ...)

\method{adsr_filter}{wave}(x, attack, decay, sustain, hold, release, ...)
}
\arguments{
\item{x}{Object to filter; will be coerced to a wave object via \code{\link{wave}}.}

\item{attack}{(Positive numeric scalar)
Duration of the attack portion, in seconds.
At the end of the attack portion, the envelope amplitude reaches 1.}

\item{decay}{(Positive numeric scalar)
Duration of the decay portion, in seconds.}

\item{sustain}{(Numeric scalar between 0 and 1)
Envelope amplitude during the sustain portion.}

\item{hold}{(Positive numeric scalar)
Duration of the sustain portion, in seconds.}

\item{release}{(Positive numeric scalar)
Duration of the release portion, in seconds.}

\item{...}{
  Arguments passed on to \code{\link[=wave]{wave}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Filters an object through the Attack-Decay-Sustain-Release (ADSR) envelope model,
implemented with linear interpolation.
}
