% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet-size.R
\name{alphabet_size}
\alias{alphabet_size}
\title{Alphabet size}
\usage{
alphabet_size(type)
}
\arguments{
\item{type}{(Character scalar) Type.}
}
\value{
Alphabet size, as an integer scalar.
}
\description{
Returns the alphabet size for a given type.
}
