% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{amplitude_to_dB}
\alias{amplitude_to_dB}
\title{Convert amplitude to decibels}
\usage{
amplitude_to_dB(amplitude, unit_amplitude_in_dB)
}
\arguments{
\item{amplitude}{Numeric vector of amplitudes, on a linear scale.}

\item{unit_amplitude_in_dB}{(Numeric scalar)
The decibel value to which an amplitude of 1 should be mapped.}
}
\value{
Numeric vector of decibels.
}
\description{
Represents an input amplitude vector in decibels.
}
\seealso{
\code{\link{dB_to_amplitude}}.
}
