% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-spectrum.R
\name{as.pc_spectrum}
\alias{as.pc_spectrum}
\title{Convert to pitch-class spectrum
Converts the input to a pitch-class spectrum.
Musical notes are converted to complex tones, after Milne & Holland (2016).}
\usage{
as.pc_spectrum(x, weights = 1, array_dim = 1200, num_harmonics = 12,
  rho = 0.75, sigma = 6.83, ...)
}
\arguments{
\item{num_harmonics}{The number of harmonics in each modelled complex tone
(including the fundamental) (numeric scalar)}

\item{rho}{The roll-off parameter determining the weight of
successive harmonics; default is 0.75, after Milne & Holland (2016)}

\item{sigma}{The standard deviation of the smoothing Gaussian;
default is 6.83, after Milne & Holland (2016)}
}
\description{
Convert to pitch-class spectrum
Converts the input to a pitch-class spectrum.
Musical notes are converted to complex tones, after Milne & Holland (2016).
}
\references{
\insertRef{Milne2016a}{hutil}
}
