% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ascending_pc_dist}
\alias{ascending_pc_dist}
\title{Get the ascending distance between two pitch classes.}
\usage{
ascending_pc_dist(x, y)
}
\arguments{
\item{x}{Vector of pitch classes}

\item{y}{Vector of pitch classes}
}
\description{
Get the ascending distance between two pitch classes.
}
