% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{cache_convert_pc_set_to_pc_spectrum}
\alias{cache_convert_pc_set_to_pc_spectrum}
\title{This function builds a comprehensive memory-based cache for
convert_pc_set_to_pc_spectrum.}
\usage{
cache_convert_pc_set_to_pc_spectrum(array_dim = 1200, num_harmonics = 12,
  rho = 0.75, sigma = 6.83)
}
\description{
This function builds a comprehensive memory-based cache for
convert_pc_set_to_pc_spectrum.
}
