% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R
\name{coded_vec}
\alias{coded_vec}
\alias{[.coded_vec}
\alias{[<-.coded_vec}
\alias{[[<-.coded_vec}
\title{Coded vector}
\usage{
coded_vec(x, type, metadata = list())

\method{[}{coded_vec}(x, i)

\method{[}{coded_vec}(x, i) <- value

\method{[[}{coded_vec}(x, i) <- value
}
\arguments{
\item{x}{(Integer vector) The coded sequence.}

\item{type}{(Character scalar) Identifies the symbol \code{\link{type}}.}

\item{metadata}{(List) A (possibly-empty) list of metadata information.}

\item{i}{(Integerish vector) Index/indices to access.}

\item{value}{New value(s).}
}
\value{
An object of class "coded_vec".
}
\description{
This function creates a coded vector.
Coded vectors are used to represent sequences of symbols
from a finite alphabet.
Each symbol should be coded as an integer using a bijective mapping.
}
\seealso{
\code{\link{vec}}.
}
