% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-spectrum.R
\name{collapse_summing_amplitudes}
\alias{collapse_summing_amplitudes}
\title{Collapse summing amplitudes}
\usage{
collapse_summing_amplitudes(x, digits, modulo = NA_real_, coherent = FALSE)
}
\arguments{
\item{x}{Input dataframe.}

\item{digits}{Number of digits to which \code{x} should be rounded.}

\item{modulo}{Optional modulo value for the rounding of \code{x}.}

\item{coherent}{Whether the amplitudes from different spectral components should be combined
assuming coherent summation, where the amplitudes simply add together
(default is \code{FALSE}).
Otherwise incoherent summation is used, where the amplitudes are squared, added, then
square rooted.}
}
\value{
A dataframe.
}
\description{
Takes a dataframe of spectral components (locations \code{x}, amplitudes \code{y}),
rounds \code{x}, and then combines spectral components with the same location.
}
