% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{concatenate}
\alias{concatenate}
\alias{concatenate.wave}
\title{Concatenate objects}
\usage{
concatenate(...)

\method{concatenate}{wave}(...)
}
\arguments{
\item{...}{Objects to combine.}
}
\description{
Concatenates multiple objects together.
}
