% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{convert_pc_set_to_pc_spectrum}
\alias{convert_pc_set_to_pc_spectrum}
\title{Convert pitch-class set to pitch-class spectrum
Takes a list of pitch classes and outputs a pitch-class spectrum obtained by treating each of the pitch classes as complex tones, after Milne & Holland (2016).}
\usage{
convert_pc_set_to_pc_spectrum(pc_set, saliences = 1, array_dim = 1200,
  num_harmonics = 12, rho = 0.75, sigma = 6.83, cache = TRUE,
  cache_env = NULL, cache_stop_on_missing = FALSE)
}
\arguments{
\item{num_harmonics}{The number of harmonics in each modelled complex tone (including the fundamental) (numeric scalar)}

\item{rho}{The roll-off parameter determining the salience of successive harmonics; default is 0.75, after Milne & Holland (2016)}

\item{sigma}{The standard deviation of the smoothing Gaussian; default is 6.83, after Milne & Holland (2016)}
}
\description{
Convert pitch-class set to pitch-class spectrum
Takes a list of pitch classes and outputs a pitch-class spectrum obtained by treating each of the pitch classes as complex tones, after Milne & Holland (2016).
}
\references{
\insertRef{Milne2016a}{HarmonyUtils}
}
