% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_pitch_to_pc}
\alias{convert_pitch_to_pc}
\title{Convert pitches to pitch classes}
\usage{
convert_pitch_to_pc(pitch)
}
\arguments{
\item{pitch}{A numeric vector of pitches}
}
\value{
A numeric vector produced by mapping each element in \code{pitch} to a pitch class
}
\description{
Converts pitches to pitch classes.
}
