% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{dB_to_amplitude}
\alias{dB_to_amplitude}
\title{Convert decibels to amplitude}
\usage{
dB_to_amplitude(dB, unit_amplitude_in_dB)
}
\arguments{
\item{dB}{Numeric vector of decibels}

\item{unit_amplitude_in_dB}{(Numeric scalar)
The decibel value that should be mapped to an amplitude of 1.}
}
\value{
Numeric vector of amplitudes, on a linear scale.
}
\description{
Represents an input vector of decibels to a linear amplitude scale.
}
\seealso{
\code{\link{amplitude_to_dB}}.
}
