% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord-qualities.R
\name{decode_chord_quality}
\alias{decode_chord_quality}
\title{Decode chord quality}
\usage{
decode_chord_quality(x, must_work = FALSE)
}
\arguments{
\item{x}{(Character scalar)
Label to decode.}

\item{must_work}{(Logical scalar)
If \code{TRUE}, the function will throw an error if the textual label
is not recognised.}
}
\value{
If the textual label is recognised,
returns an integer vector corresponding to the pitch-class set
expressed relative to the chord root,
otherwise returns \code{NULL}
(or an error if \code{must_work} is \code{TRUE}).
}
\description{
Decodes a given textual label into the implied chord quality.
}
\details{
No preprocessing is applied to \code{x}.
}
\seealso{
\itemize{
\item \code{\link{register_chord_quality}} for registering additional chord qualities
in the database.
\item \code{\link{initialise_chord_qualities}} for resetting the database.
}
}
