% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord_alphabet.R
\name{decode_chords}
\alias{decode_chords}
\title{Decode chords}
\usage{
decode_chords(chords)
}
\value{
List of decoded chords.
}
\description{
This function is faster than repeated application of <decode_chord>.
}
