% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{description}
\alias{description}
\title{#' @export
get_transpositions <- function(x) UseMethod("get_transpositions")
#' @export
normalise_bass <- function(x) UseMethod("normalise_bass")}
\usage{
description(x)
}
\description{
#' @export
get_transpositions <- function(x) UseMethod("get_transpositions")
#' @export
normalise_bass <- function(x) UseMethod("normalise_bass")
}
