% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-chord.R
\name{.fr_chord}
\alias{.fr_chord}
\title{Frequency chord constructor}
\usage{
.fr_chord(...)
}
\arguments{
\item{x}{(Numeric vector) Frequencies in ascending order}
}
\description{
This hidden function constructs a frequency chord object.
It is unforgiving with respect to input formats,
unlike \code{\link{fr_chord}}.
}
\keyword{internal}
