% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord.R
\name{.pc_chord}
\alias{.pc_chord}
\title{Pitch-class chord constructor}
\usage{
.pc_chord(bass_pc, other_pc = numeric())
}
\arguments{
\item{bass_pc}{Numeric scalar corresponding to bass pitch class}

\item{other_pc}{Numeric vector corresponding to pitch-class set,
may optionally include the bass pitch class}
}
\description{
This hidden function constructs a pitch-class chord object.
It is unforgiving with respect to input formats,
unlike \code{\link{pc_chord}}.
}
\keyword{internal}
