% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord-type.R
\name{.pc_chord_type}
\alias{.pc_chord_type}
\title{Pitch-class chord type constructor}
\usage{
.pc_chord_type(x)
}
\description{
This is the low-level construct for pitch-class chord type.
}
\keyword{internal}
