% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-milne-spectrum.R
\name{.pc_milne_spectrum}
\alias{.pc_milne_spectrum}
\title{Constructor function for pitch-class Milne spectrum}
\usage{
.pc_milne_spectrum(x)
}
\arguments{
\item{x}{A numeric vector of pitch-class weights,
typically (but not necessarily) of length 1200.}
}
\value{
An object of class "pc_milne_spectrum".
}
\description{
This function constructs a "pc_milne_spectrum" object.
}
\seealso{
\code{\link{pc_milne_spectrum}}.
}
