% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{.pi_chord}
\alias{.pi_chord}
\title{pi_chord constructor}
\usage{
.pi_chord(...)
}
\arguments{
\item{x}{(Numeric vector) MIDI note numbers in ascending order}
}
\description{
Constructor function for objects of class "pi_chord".
}
\keyword{internal}
