% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{.wave}
\alias{.wave}
\title{wave constructor}
\usage{
.wave(x, sample_rate)
}
\arguments{
\item{x}{Numeric vector of displacement values.}

\item{sample_rate}{(Numeric scalar) Sample rate.}
}
\description{
Constructor function for objects of class "wave".
}
