% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord.R
\name{edit_bass_pc}
\alias{edit_bass_pc}
\alias{edit_bass_pc.pc_chord}
\title{Edit bass pitch class}
\usage{
edit_bass_pc(x, new)

\method{edit_bass_pc}{pc_chord}(x, new)
}
\arguments{
\item{x}{Original sonority.}

\item{new}{(Numeric scalar)
New bass pitch class.}
}
\value{
A sonority with the updated bass pitch class.
}
\description{
Edits the bass pitch class of a sonority.
Throws an error if the proposed pitch class was not already
in the sonority's pitch-class set.
}
