% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R, R/pc-chord.R,
%   R/pc-set-norm-form.R, R/pc-set.R
\name{encode}
\alias{encode}
\alias{encode.vec}
\alias{encode.coded_vec}
\alias{encode.corpus}
\alias{encode.pc_chord}
\alias{encode.pc_set_norm_form}
\alias{encode.pc_set}
\title{Encode}
\usage{
encode(x)

\method{encode}{vec}(x)

\method{encode}{coded_vec}(x)

\method{encode}{corpus}(x)

\method{encode}{pc_chord}(x)

\method{encode}{pc_set_norm_form}(x)

\method{encode}{pc_set}(x)
}
\arguments{
\item{x}{Object to transform.}
}
\value{
Encoded object.
}
\description{
Transforms a given object into an integer-based encoding.
}
\details{
Encoding is currently defined for the following types:
\itemize{
\item \code{\link{pc_set}}
\item \code{\link{pc_chord}}
\item \code{\link{pc_chord_type}}
}

An error will be thrown when trying to encode objects of other types.
}
\seealso{
\code{\link{decode}} for the reverse operation.
}
