% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-harmonics.R
\name{expand_harmonics.fr_sparse_spectrum}
\alias{expand_harmonics.fr_sparse_spectrum}
\title{Expand harmonics}
\usage{
\method{expand_harmonics}{fr_sparse_spectrum}(x, num_harmonics = 11,
  roll_off = 1, frequency_digits = 6, ...)
}
\arguments{
\item{num_harmonics}{(Integerish scalar) Number of harmonics produced by each tone,
including the fundamental frequency.}

\item{roll_off}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}

\item{frequency_digits}{(Integerish scalar) Number of significant digits
to which frequencies are rounded before adjacent partials are combined.}
}
\description{
Expands each tone in an object into its implied harmonics.
}
