% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-harmonics.R
\name{expand_harmonics}
\alias{expand_harmonics}
\alias{expand_harmonics.sparse_fr_spectrum}
\alias{expand_harmonics.sparse_pi_spectrum}
\alias{expand_harmonics.pi_chord}
\title{Expand harmonics}
\usage{
expand_harmonics(x, num_harmonics = 11L, roll_off = 1, digits = 6)

\method{expand_harmonics}{sparse_fr_spectrum}(x, num_harmonics = 11L,
  roll_off = 1, digits = 6)

\method{expand_harmonics}{sparse_pi_spectrum}(x, num_harmonics = 11L,
  roll_off = 1, digits = 6)

\method{expand_harmonics}{pi_chord}(x, ...)
}
\arguments{
\item{x}{Object whose harmonics should be expanded.}

\item{num_harmonics}{(Integerish scalar) Number of harmonics produced by each tone,
including the fundamental frequency.}

\item{roll_off}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}

\item{...}{Further parameters to be passed to individual methods.}

\item{round}{Whether or not the harmonic template should be rounded
to the nearest chromatic pitch.}
}
\description{
Expands each tone in an object into its implied harmonics.
}
