% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-filter.R
\name{filter_spectrum_gaussian}
\alias{filter_spectrum_gaussian}
\title{Gaussian filter}
\usage{
filter_spectrum_gaussian(x, location, width = 50, ...)
}
\arguments{
\item{x}{Object to filter, passed to \code{\link{sparse_fr_spectrum}}.}

\item{location}{Vector of locations for the Gaussians.}

\item{width}{Vector of widths (i.e. SDs) for the Gaussians.
If only one number is provided, this same number is used for all Gaussians.}

\item{...}{Optional parameters passed to \code{\link{sparse_fr_spectrum}}.}
}
\description{
Applies a Gaussian filter to a sparse spectrum.
}
