% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-sparse-spectrum.R
\name{fr_sparse_spectrum}
\alias{fr_sparse_spectrum}
\alias{fr_sparse_spectrum.pi_sparse_spectrum}
\alias{fr_sparse_spectrum.pi_chord}
\alias{fr_sparse_spectrum.default}
\alias{fr_sparse_spectrum.list}
\title{Frequency sparse spectrum}
\usage{
fr_sparse_spectrum(x, ...)

\method{fr_sparse_spectrum}{pi_sparse_spectrum}(x, ...)

\method{fr_sparse_spectrum}{pi_chord}(x, ...)

\method{fr_sparse_spectrum}{default}(x, ...)

\method{fr_sparse_spectrum}{list}(x, ...)
}
\arguments{
\item{x}{Input sonority.
\itemize{
\item Numeric vectors will be treated as vectors of MIDI note numbers,
and expanded into their implied harmonics.
\item Two-element lists will be treated as finalised spectra,
with the first element being a numeric vector of frequencies,
and the second element being a numeric vector of amplitudes.
}}

\item{...}{Further arguments passed to \code{\link{expand_harmonics}()},
depending on the method invoked.}
}
\value{
An object of class \code{fr_sparse_spectrum}.
}
\description{
This function represents an input object as a
'frequency sparse spectrum'.
}
\details{
A frequency sparse spectrum comprises a finite set of spectral components,
each defined by a frequency (in Hz)
and an amplitude (expressed in arbitrary units, but with the
fundamental frequencies of chord pitches typically taking the value 1).
}
