% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{freq_to_midi}
\alias{freq_to_midi}
\title{Convert frequencies to MIDI note numbers}
\usage{
freq_to_midi(frequency,
  stretched_octave = get_midi_params()$stretched_octave,
  tuning_ref_Hz = get_midi_params()$tuning_ref_Hz)
}
\arguments{
\item{stretched_octave}{Logical scalar; whether or not to use a stretched octave. Default is \code{FALSE}}

\item{tuning_ref_Hz}{The tuning reference point in Hz, should correspond to the frequency of the A above middle C (typically 440 Hz)}

\item{midi}{Numeric vector of MIDI note numbers}
}
\value{
Numeric vector of frequencies in Hz
}
\description{
Converts frequencies (Hz) to MIDI note numbers, optionally using stretched octaves. Non-integer MIDI note numbers are permitted as output.
}
