% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-chord.R
\name{get_bass_fr}
\alias{get_bass_fr}
\alias{get_bass_fr.default}
\alias{get_bass_fr.fr_chord}
\title{Get bass frequency}
\usage{
get_bass_fr(x)

\method{get_bass_fr}{default}(x)

\method{get_bass_fr}{fr_chord}(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
Bass frequency, as a numeric scalar.
}
\description{
Gets the bass frequency from a sonority.
}
