% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{get_bass_pi}
\alias{get_bass_pi}
\alias{get_bass_pi.default}
\alias{get_bass_pi.pi_chord}
\title{Get bass pitch}
\usage{
get_bass_pi(x)

\method{get_bass_pi}{default}(x)

\method{get_bass_pi}{pi_chord}(x)
}
\arguments{
\item{x}{Input sonority.}
}
\value{
The bass pitch, as a MIDI note number (numeric scalar).
}
\description{
Gets the bass pitch of a sonority.
}
