% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord-qualities.R
\name{get_chord_qualities_df}
\alias{get_chord_qualities_df}
\title{Get chord qualities}
\usage{
get_chord_qualities_df()
}
\description{
Returns a \code{\link[tibble]{tibble}} of chord qualities and different
textual labels for these chord qualities.
In practice most users will access this information through the
function \code{\link{decode_chord_quality}}.
}
\details{
The \code{tibble} has one row for each unique chord quality,
where a chord quality is defined as a set of pitch classes
expressed relative to the chord root.
\itemize{
\item The \code{pitch_classes} column provides the chord quality as an integer vector,
corresponding to a list of pitch classes expressed relative to the chord root.
\item The \code{scale_degrees} column lists the scale degrees present in the
chord quality in a more traditional human-readable format.
\item The \code{description} column provides a textual description for the
chord quality.
\item The remaining columns are prefixed by \code{key_},
and provide textual abbreviations for the chord quality.
These \code{key_} elements define the legal inputs for the function
\code{\link{decode_chord_quality}}.
}

The data is read from the package file 'inst/chord-qualities.csv'.
Additions can be proposed by GitHub: \url{https://github.com/pmcharrison/hrep}.
}
\seealso{
\itemize{
\item \code{\link{decode_chord_quality}} for decoding chord qualities.
\item \code{\link{register_chord_quality}} for registering additional chord qualities.
}
}
