% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{get_complex_tone}
\alias{get_complex_tone}
\title{Get complex tone}
\usage{
get_complex_tone(fundamental_pc, array_dim, num_harmonics, rho, sigma)
}
\description{
Wrapper for \code{make_complex_tone} that implements caching.
}
