% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_harmonic_template}
\alias{get_harmonic_template}
\title{Get harmonic template}
\usage{
get_harmonic_template(num_harmonics, amplitude, roll_off = 1)
}
\arguments{
\item{num_harmonics}{Number of harmonics (including fundamental)}

\item{amplitude}{amplitude of the fundamental frequency}
}
\value{
\code{data.frame} of frequencies and amplitudes
}
\description{
Gets a harmonic template, using non-stretched octaves.
}
