% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_harmonic_template_midi}
\alias{get_harmonic_template_midi}
\title{Get harmonic template}
\usage{
get_harmonic_template_midi(num_harmonics, level, roll_off = function(x) 1/(1 +
  x))
}
\arguments{
\item{num_harmonics}{Number of harmonics (including fundamental)}

\item{level}{Level of the fundamental frequency}

\item{roll_off}{Function determining the level of the nth harmonic. Default value corresponds to Equation 9 of Parncutt and Strasburger (1994)}
}
\description{
Gets a harmonic template.
}
