% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord.R
\name{get_non_bass_pc}
\alias{get_non_bass_pc}
\title{Get non-bass pitch classes}
\usage{
get_non_bass_pc(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
The non-bass pitch classes, as a numeric vector.
}
\description{
Gets the non-bass pitch classes in a sonority.
}
