% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_pc_dist}
\alias{get_pc_dist}
\title{Get the distance between two pitch classes.}
\usage{
get_pc_dist(x, y)
}
\arguments{
\item{x}{Vector of pitch classes}

\item{y}{Vector of pitch classes}
}
\description{
Get the distance between two pitch classes.
}
