% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pc_set_alphabet}
\alias{get_pc_set_alphabet}
\title{Get pitch class set alphabet}
\usage{
get_pc_set_alphabet()
}
\value{
List of all possible pitch class sets.
}
\description{
Returns a list of all possible pitch class sets.
}
