% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrep-package.R
\docType{package}
\name{hrep-package}
\alias{hrep}
\alias{hrep-package}
\title{hrep: Low-level utility functions for harmony analysis packages}
\description{
Provides low-level utility functions for harmony analysis packages. In most music analysis situations you will probably want the package HarmonyToolbox instead.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
