% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord-qualities.R
\name{initialise_chord_qualities}
\alias{initialise_chord_qualities}
\title{Initialise chord qualities database}
\usage{
initialise_chord_qualities()
}
\description{
(Re)initialises the chord qualities used by the \code{\link{decode_chord_quality}}
function to values specified by the \code{\link{get_chord_qualities_df}}
function.
}
