% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int-vec.R
\name{int_vec}
\alias{int_vec}
\alias{int_vec.pc_set}
\alias{int_vec.default}
\alias{as.numeric.int_vec}
\alias{as.character.int_vec}
\title{Interval vector}
\usage{
int_vec(x)

\method{int_vec}{pc_set}(x)

\method{int_vec}{default}(x)

\method{as.numeric}{int_vec}(x, ...)

\method{as.character}{int_vec}(x, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{...}{Further arguments passed to other methods.}
}
\value{
An integer vector of length 6,
where the ith element indicates the number of times that
the pitch-class interval i is found in the chord.
}
\description{
Computes a chord's interval vector, after
\insertCite{Parncutt2018;textual}{hrep}.
}
\details{
An interval vector of class "int_vec",
describing how often each pitch-class interval
appears in a chord.
There are six possible interval classes: 1, 2, 3, 4, 5, and 6 semitones.
\insertCite{Parncutt2018;textual}{hrep} cite
\insertCite{Forte1977;textual}{hrep} for this technique.
}
\examples{
int_vec(c(60, 64, 67)) # major triad
int_vec(c(60, 63, 66)) # diminished triad
}
\references{
\insertAllCited{}
}
