% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R, R/generics.R
\name{is.coded.coded_vec}
\alias{is.coded.coded_vec}
\alias{is.coded.corpus}
\alias{is.coded}
\title{Is an object coded?}
\usage{
\method{is.coded}{coded_vec}(x)

\method{is.coded}{corpus}(x)

is.coded(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Scalar logical.
}
\description{
Checks whether an object is integer-coded.
}
\seealso{
\code{\link{encode}}, \code{\link{decode}}
}
