% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R
\name{is.coded_vec}
\alias{is.coded_vec}
\title{Type-checking for "coded_vec"}
\usage{
is.coded_vec(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of type "coded_vec".
}
