% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.equal-tempered.R
\name{is.equal_tempered}
\alias{is.equal_tempered}
\alias{is.equal_tempered.pi_chord}
\alias{is.equal_tempered.pc_set}
\alias{is.equal_tempered.pc_chord}
\title{Check for equal temperament}
\usage{
is.equal_tempered(x, ...)

\method{is.equal_tempered}{pi_chord}(x, ...)

\method{is.equal_tempered}{pc_set}(x, ...)

\method{is.equal_tempered}{pc_chord}(x, ...)
}
\arguments{
\item{x}{Chord to check.}

\item{...}{Further arguments passed to other methods.}
}
\value{
Returns \code{TRUE} if the chord is equal-tempered,
\code{FALSE} otherwise.
}
\description{
Checks whether a chord is equal-tempered.
}
