% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milne-pc-spectrum.R
\name{is.milne_pc_spectrum}
\alias{is.milne_pc_spectrum}
\title{Check for class "milne_pc_spectrum"}
\usage{
is.milne_pc_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class "milne_pc_spectrum".
}
