% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-milne-spectrum.R
\name{is.pc_milne_spectrum}
\alias{is.pc_milne_spectrum}
\title{Check for class "pc_milne_spectrum"}
\usage{
is.pc_milne_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class "pc_milne_spectrum".
}
