% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set.R
\name{is.pc_set}
\alias{is.pc_set}
\title{Check for type "pc_set"}
\usage{
is.pc_set(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Scalar logical.
}
\description{
Checks whether an object is of type "pc_set".
}
