% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{is.pi_chord}
\alias{is.pi_chord}
\title{Check for type "pi_chord"}
\usage{
is.pi_chord(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of type "pi_chord".
}
