% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-pc-spectrum.R
\name{is.smooth_pc_spectrum}
\alias{is.smooth_pc_spectrum}
\title{Is smooth pitch-class spectrum}
\usage{
is.smooth_pc_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object belongs to the class \code{smooth_pc_spectrum}.
}
