% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-pi-spectrum.R
\name{is.smooth_pi_spectrum}
\alias{is.smooth_pi_spectrum}
\title{Is smooth pitch spectrum}
\usage{
is.smooth_pi_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object belongs to the class \code{smooth_pi_spectrum}.
}
