% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-fr-spectrum.R
\name{is.sparse_fr_spectrum}
\alias{is.sparse_fr_spectrum}
\title{Is sparse pitch-class spectrum}
\usage{
is.sparse_fr_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object belongs to the class \code{sparse_fr_spectrum}.
}
