% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-pi-spectrum.R
\name{is.sparse_pi_spectrum}
\alias{is.sparse_pi_spectrum}
\title{Is sparse pitch spectrum}
\usage{
is.sparse_pi_spectrum(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object belongs to the class \code{sparse_pi_spectrum}.
}
