% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-chords.R
\name{list_chords}
\alias{list_chords}
\title{List chords}
\usage{
list_chords(type)
}
\arguments{
\item{type}{(Character scalar) Type to list.}
}
\value{
A list of all chords of the requested type,
where the ith element corresponds to the ith element of the alphabet.
}
\description{
Lists all chords of a given type.
}
