% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{make_gaussian_spectral_template}
\alias{make_gaussian_spectral_template}
\title{Make Gaussian spectral template}
\usage{
make_gaussian_spectral_template(array_dim, sigma, truncation_point = 12)
}
\description{
Makes a Gaussian spectral template with unit mass, centred on 0, with standard deviation <sigma>. The template will be truncated to zero for points <truncation-point> standard deviations or further away from the mean, after Milne's implementation.
}
