% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{metadata}
\alias{metadata}
\alias{metadata<-}
\title{Metadata}
\usage{
metadata(x)

metadata(x) <- value
}
\arguments{
\item{x}{Method whose metadata should be accessed.}

\item{value}{New value.}
}
\description{
Additional information about an object may be stored in its
metadata attribute.
}
