% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{midi_to_freq}
\alias{midi_to_freq}
\title{Convert MIDI note numbers to frequencies}
\usage{
midi_to_freq(midi, stretched_octave = get_midi_params()$stretched_octave,
  tuning_ref_Hz = get_midi_params()$tuning_ref_Hz)
}
\arguments{
\item{midi}{Numeric vector of MIDI note numbers}

\item{stretched_octave}{Logical scalar; whether or not to use a stretched octave. Default is \code{FALSE}}

\item{tuning_ref_Hz}{The tuning reference point in Hz, should correspond to the frequency of the A above middle C (typically 440 Hz)}
}
\value{
Numeric vector of frequencies in Hz
}
\description{
Converts MIDI note numbers to frequencies (Hz), optionally using stretched octaves. Corresponds to Equation 1 of Parncutt & Strasburger (1994), but with Hz instead of kHz.
}
