% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milne-pc-spec-dist.R
\name{milne_pc_spec_dist}
\alias{milne_pc_spec_dist}
\title{Milne's pitch-class spectral distance}
\usage{
milne_pc_spec_dist(x, y, ...)
}
\arguments{
\item{x}{First sonority, passed to \code{\link{milne_pc_spectrum}}.}

\item{y}{Second sonority, passed to \code{\link{milne_pc_spectrum}}.}

\item{...}{Further arguments passed to \code{\link{milne_pc_spectrum}}.}
}
\value{
A numeric scalar where greater values indicate greater distance.
0 is the minimum possible distance.
1 corresponds to two uncorrelated spectra;
values greater than 1 are possible if the spectra are inversely correlated.
}
\description{
Computes Milne's pitch-class spectral distance measure between two sonorities.
}
\details{
See \insertCite{Milne2016;textual}{hrep} for details.
}
\references{
\insertAllCited{}
}
