% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milne-pc-spectrum.R
\name{milne_pc_spectrum}
\alias{milne_pc_spectrum}
\alias{milne_pc_spectrum.pc_set}
\alias{milne_pc_spectrum.default}
\alias{milne_pc_spectrum.milne_pc_spectrum}
\title{Milne pitch-class spectrum}
\usage{
milne_pc_spectrum(x, ...)

\method{milne_pc_spectrum}{pc_set}(
  x,
  weights = 1,
  num_harmonics = 12,
  rho = 0.75,
  sigma = 6.83,
  array_dim = 1200,
  ...
)

\method{milne_pc_spectrum}{default}(x, ...)

\method{milne_pc_spectrum}{milne_pc_spectrum}(x, ...)
}
\arguments{
\item{x}{Input sonority.}

\item{...}{Provided for S3 method consistency.}

\item{weights}{(Numeric vector)
Vector of weights to assign to each pitch class.
If a scalar value is provided, this value is assigned to all pitch classes.}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics to use when expanding tones into their implied harmonics,
and when defining the harmonic template
(including the fundamental frequency).
Defaults to 12, after
\insertCite{Milne2016;textual}{hrep}.}

\item{rho}{(Numeric scalar)
Roll-off parameter for harmonic expansion.
Defaults to 0.75, after
\insertCite{Milne2016;textual}{hrep}.}

\item{sigma}{(Numeric scalar)
Standard deviation of the Gaussian smoothing distribution (cents).
Defaults to 6.83, after
\insertCite{Milne2016;textual}{hrep}.}

\item{array_dim}{(Integerish scalar)
Dimensionality of the pitch-class spectrum array.
Defaults to 1200, after
\insertCite{Milne2016;textual}{hrep}.}
}
\value{
An object of class \code{milne_pc_spectrum}.
}
\description{
This function represents an input object as a
'Milne pitch-class spectrum'.
A Milne pitch-class spectrum defines 'perceptual weight'
as a continuous function of 'pitch class'.
}
\details{
This spectrum is typically constructed from musical chords
by expanding each note into its implied harmonics
and applying a Gaussian smoothing to account for perceptual uncertainties.
See \insertCite{Milne2016;textual}{hrep} for details.
}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link{.milne_pc_spectrum}}.
}
