% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-smooth-spectrum.R
\name{new_complex_tone}
\alias{new_complex_tone}
\title{Make complex tone}
\usage{
new_complex_tone(fundamental_pc, weight, array_dim, num_harmonics, rho,
  sigma)
}
\arguments{
\item{num_harmonics}{Number of harmonics, including the fundamental}
}
\description{
Returns an array describing the pitch-class spectrum for a given complex tone.
}
