% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R, R/generics.R, R/vec.R
\name{num_elements.coded_vec}
\alias{num_elements.coded_vec}
\alias{num_elements.corpus}
\alias{num_elements}
\alias{num_elements.vec}
\title{Number of elements}
\usage{
\method{num_elements}{coded_vec}(x)

\method{num_elements}{corpus}(x)

num_elements(x)

\method{num_elements}{vec}(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
The size of \code{x}, as an integer scalar.
}
\description{
Returns the number of elements in an object.
For an object of class \code{\link{vec}}, this will be the length of the vector.
For an object of class \code{\link{corpus}},
this will be the sum of the lengths of its constituent sequences.
}
