% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R, R/generics.R
\name{num_sequences.corpus}
\alias{num_sequences.corpus}
\alias{num_sequences}
\title{Number of sequences}
\usage{
\method{num_sequences}{corpus}(x)

num_sequences(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
The number of sequences in \code{x}, as an integer scalar.
}
\description{
Counts the number of sequences in an object.
}
