% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{pad}
\alias{pad}
\title{Pad}
\usage{
pad(x, before, after)
}
\arguments{
\item{before}{(Numeric scalar)
Seconds of silence to add before the sound.}

\item{after}{(Numeric scalar)
Seconds of silence to add after the sound.}
}
\description{
Adds silent padding to the beginning/end of a sound.
}
