% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord.R
\name{pc_chord}
\alias{pc_chord}
\alias{pc_chord.numeric}
\alias{pc_chord.pc_set}
\alias{pc_chord.pc_chord}
\alias{pc_chord.pc_chord_type}
\alias{pc_chord.pi_chord}
\alias{pc_chord.fr_chord}
\title{Pitch-class chord}
\usage{
pc_chord(x)

\method{pc_chord}{numeric}(x)

\method{pc_chord}{pc_set}(x)

\method{pc_chord}{pc_chord}(x)

\method{pc_chord}{pc_chord_type}(x)

\method{pc_chord}{pi_chord}(x)

\method{pc_chord}{fr_chord}(x)
}
\arguments{
\item{x}{Object to represent as a pitch-class chord.}
}
\value{
Returns an object of class \code{pc_chord}.
}
\description{
This function represents an object as a pitch-class chord.
A pitch-class chord is defined by the combination of
a pitch-class set and a bass pitch class.
}
