% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_chord_id_to_pc_set_id_map}
\alias{pc_chord_id_to_pc_set_id_map}
\title{pc_chord to pc_set map}
\description{
This integer vector provides the mapping between
pitch-class chords ("pc_chord" objects) and pitch-class sets ("pc_set" objects).
This can help computational efficiency in some situations.
}
\details{
If \code{i} is the integer corresponding to an encoded pitch-class chord,
then the corresponding encoded pitch-class set is
equal to the ith element of this vector.
}
\examples{
i <- encode(pc_chord(c(4, 0, 7)))
j <- pc_chord_id_to_pc_set_id_map[i]
decode(coded_vec(j, "pc_set"))[[1]]
}
\keyword{data}
