% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord-type.R, R/pi-chord-type.R
\name{pc_chord_type}
\alias{pc_chord_type}
\alias{pc_chord_type.default}
\alias{pc_chord_type.pc_chord}
\alias{pc_chord_type.pi_chord}
\alias{pc_chord_type.pc_set}
\alias{pc_chord_type.pc_set_type}
\alias{pc_chord_type.smooth_spectrum}
\alias{pc_chord_type.sparse_spectrum}
\alias{pi_chord_type.pi_chord}
\alias{pi_chord_type.smooth_spectrum}
\alias{pi_chord_type.sparse_spectrum}
\title{Pitch-class chord type}
\usage{
pc_chord_type(x)

\method{pc_chord_type}{default}(x)

\method{pc_chord_type}{pc_chord}(x)

\method{pc_chord_type}{pi_chord}(x)

\method{pc_chord_type}{pc_set}(x)

\method{pc_chord_type}{pc_set_type}(x)

\method{pc_chord_type}{smooth_spectrum}(x)

\method{pc_chord_type}{sparse_spectrum}(x)

\method{pi_chord_type}{pi_chord}(x)

\method{pi_chord_type}{smooth_spectrum}(x)

\method{pi_chord_type}{sparse_spectrum}(x)
}
\arguments{
\item{x}{Object to represent as a pitch-class chord type.}
}
\value{
Returns an object of class \code{pc_chord_type}.
}
\description{
This function represents an object as a pitch-class chord type.
A pitch-class chord type is defined as the chord's pitch class set
expressed relative to the pitch class of the bass note.
}
