% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-milne-spectrum.R
\name{pc_milne_spectrum}
\alias{pc_milne_spectrum}
\alias{pc_milne_spectrum.pc_set}
\alias{pc_milne_spectrum.default}
\title{Pitch-class Milne spectrum}
\usage{
pc_milne_spectrum(x, ...)

\method{pc_milne_spectrum}{pc_set}(x, weights = 1, num_harmonics = 12,
  rho = 0.75, sigma = 6.83, array_dim = 1200, ...)

\method{pc_milne_spectrum}{default}(x, ...)
}
\arguments{
\item{x}{Input sonority.}

\item{...}{Further arguments passed to specific methods.}

\item{weights}{(Numeric vector)
Vector of weights to assign to each pitch class.
If a scalar value is provided, this value is assigned to all pitch classes.}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics to use when expanding tones into their implied harmonics,
and when defining the harmonic template
(including the fundamental frequency).
Defaults to 12, after
\insertCite{Milne2016;textual}{har18}.}

\item{rho}{(Numeric scalar)
Roll-off parameter for harmonic expansion.
Defaults to 0.75, after
\insertCite{Milne2016;textual}{har18}.}

\item{sigma}{(Numeric scalar)
Standard deviation of the Gaussian smoothing distribution (cents).
Defaults to 6.83, after
\insertCite{Milne2016;textual}{har18}.}

\item{array_dim}{(Integerish scalar)
Dimensionality of the pitch-class spectrum array.
Defaults to 1200, after
\insertCite{Milne2016;textual}{har18}.}
}
\value{
An object of class \code{pc_milne_spectrum}.
}
\description{
This function represents an input object as a
'pitch-class Milne spectrum'.
A pitch-class Milne spectrum defines 'perceptual weight'
as a continuous function of 'pitch class'.
}
\details{
This spectrum is typically constructed from musical chords
by expanding each note into its implied harmonics
and applying a Gaussian smoothing to account for perceptual uncertainties.
See \insertCite{Milne2016a;textual}{hrep} for details.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{.pc_milne_spectrum}}.
}
