% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set-norm-form.R
\name{pc_set_norm_form}
\alias{pc_set_norm_form}
\alias{pc_set_norm_form.default}
\alias{pc_set_norm_form.pc_set}
\alias{pc_set_norm_form.pc_set_norm_order}
\alias{pc_set_norm_form.pc_set_norm_form}
\alias{transposition.pc_set_norm_form}
\title{Pitch-class set (normal form)}
\usage{
pc_set_norm_form(x)

\method{pc_set_norm_form}{default}(x)

\method{pc_set_norm_form}{pc_set}(x)

\method{pc_set_norm_form}{pc_set_norm_order}(x)

\method{pc_set_norm_form}{pc_set_norm_form}(x)

\method{transposition}{pc_set_norm_form}(x)
}
\arguments{
\item{x}{Input sonority.}
}
\value{
An object of class "pc_set_norm_form".
}
\description{
Expresses a sonority as a normal-form pitch-class set.
Normal-form pitch-class sets form equivalence classes of
pitch-class sets under transposition.
}
\note{
When created from a pitch-class set, a normal-form pitch-class set
is associated with the interval by which it was transposed
from the original pitch-class set.
This transposition can be accessed with the function \code{transposition()}.
}
