% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set-norm-order.R
\name{pc_set_norm_order}
\alias{pc_set_norm_order}
\alias{pc_set_norm_order.numeric}
\alias{pc_set_norm_order.pc_set}
\title{Pitch-class set (normal order)}
\usage{
pc_set_norm_order(x)

\method{pc_set_norm_order}{numeric}(x)

\method{pc_set_norm_order}{pc_set}(x)
}
\arguments{
\item{x}{Input sonority.}
}
\value{
An object of class "pc_set_norm_order"
}
\description{
The normal order of a pitch-class set is a maximally compact
reordering of the original pitch-class set.
This function expresses an input sonority
as a normal-order pitch-class set.
}
