% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set-norm-form.R
\name{pc_set_type}
\alias{pc_set_type}
\alias{pc_set_type.default}
\alias{pc_set_type.pc_set}
\alias{pc_set_type.pc_set_norm_order}
\alias{pc_set_type.pc_set_type}
\alias{transposition.pc_set_type}
\title{Pitch-class set type}
\usage{
pc_set_type(x)

\method{pc_set_type}{default}(x)

\method{pc_set_type}{pc_set}(x)

\method{pc_set_type}{pc_set_norm_order}(x)

\method{pc_set_type}{pc_set_type}(x)

\method{transposition}{pc_set_type}(x)
}
\arguments{
\item{x}{Input sonority.}
}
\value{
An object of class "pc_set_type".
}
\description{
Expresses a sonority as a pitch-class set type.
Pitch-class set types form equivalence classes of
pitch-class sets under transposition.
}
\note{
When created from a pitch-class set, a pitch-class set type
is associated with the interval by which it was transposed
from the original pitch-class set.
This transposition can be accessed with the function \code{transposition()}.
}
