% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_set_type_alphabet}
\alias{pc_set_type_alphabet}
\title{Pitch-class set type alphabet}
\description{
This dataset defines an alphabet for the type "pc_set_type".
It is a list with two components,
defining a forward and backward
mapping between integers and "pc_set_type" objects:
\itemize{
\item \code{by_id}:
A list object.
The ith element of this list contains the pitch-class set type
corresponding to integer i.
\item \code{by_chord}:
An environment object.
A pitch-class set type can be mapped to an integer by
taking the pitch-class set type object,
converting it to a string representation (with \code{as.character}),
and accessing the object within this environment with this name.
The result will be a coded vector (see \code{\link{coded_vec}})
of length 1.
}
}
\keyword{data}
