% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord-type.R
\name{pi_chord_type}
\alias{pi_chord_type}
\alias{pi_chord_type.default}
\title{Pitch chord type}
\usage{
pi_chord_type(x)

\method{pi_chord_type}{default}(x)
}
\arguments{
\item{x}{Object to represent as a pitch chord type.}
}
\value{
Returns an object of class \code{pi_chord_type}.
}
\description{
This function represents an object as a pitch chord type.
A pitch chord type is defined as the chord's pitches
expressed relative to the bass pitch.
}
