% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{pi_to_pc}
\alias{pi_to_pc}
\title{Convert pitches to pitch classes}
\usage{
pi_to_pc(x)
}
\arguments{
\item{x}{A numeric vector of pitches}
}
\value{
A numeric vector produced by mapping each element in \code{pitch} to a pitch class
}
\description{
Converts pitches to pitch classes.
}
