% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pitch_to_pc_set}
\alias{pitch_to_pc_set}
\title{Convert pitches to pitch class set}
\usage{
pitch_to_pc_set(pitch)
}
\arguments{
\item{pitch}{A numeric vector of pitches, possibly including duplicates}
}
\value{
A numeric vector of pitch classes, produced by mapping each element in \code{pitch} to a pitch class, sorting, and removing duplicates.
}
\description{
Converts a vector of pitches to a pitch class set.
}
