% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-pluck.R
\name{play_pluck}
\alias{play_pluck}
\title{Play sound (pluck)}
\usage{
play_pluck(x, ...)
}
\arguments{
\item{x}{Object to play (see \code{\link{save_wav_pluck}} for valid options).}

\item{...}{Further parameters to pass to \code{\link{save_wav_pluck}}.}
}
\description{
Plays a sound with a plucked timbre.
}
\details{
The sound is synthesised using \code{\link{save_wav_pluck}}
and saved to a temporary file, which is then played from the R session.
}
\note{
The command-line sound-processing program sox
(\url{http://sox.sourceforge.net/})
must be installed and available on the command line,
making available the commands \code{sox} and \code{play}.
}
