% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-sox.R
\name{play_sox}
\alias{play_sox}
\title{Play sound (sox)}
\usage{
play_sox(x, ...)
}
\arguments{
\item{x}{Object to play (see \code{\link{save_wav_sox}} for valid options).}

\item{...}{Further parameters to pass to \code{\link{save_wav_sox}}.}
}
\description{
Plays a sound using the command-line tool sox.
}
\details{
The sound is synthesised using \code{\link{save_wav_sox}}
and saved to a temporary file, which is then played from the R session.
}
\note{
The command-line sound-processing program sox
(\url{http://sox.sourceforge.net/})
must be installed and available on the command line,
making available the commands \code{sox} and \code{play}.
}
