% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{play_wav}
\alias{play_wav}
\title{Play wav}
\usage{
play_wav(x, player = "play", ...)
}
\arguments{
\item{x}{Chord to save.}

\item{player}{(Path to) a program capable of playing a wave file by invocation from the command line.
        If under Windows and no player is given, \dQuote{mplay32.exe} or \dQuote{wmplayer.exe} 
        (if the former does not exists as under Windows 7) will be chosen as the default.}

\item{...}{
  Arguments passed on to \code{\link[=save_wav]{save_wav}}, \code{\link[=expand_harmonics]{expand_harmonics}}
  \describe{
    \item{\code{file}}{(Character scalar) Output file.}
    \item{\code{amplitude}}{(Numeric scalar)
The wave is multiplied by this number before exporting.
The resulting wave should fall completely within the range [-1, 1].}
    \item{\code{bit_depth}}{(Integer scalar)
The bit depth of the exported audio.}
    \item{\code{end_pad}}{(Numeric scalar)
Duration of silence (seconds) appended to the end of the audio file,
used to avoid clicks and other artifacts.}
    \item{\code{length_sec}}{(Numeric scalar) Length of the output wave, in seconds.}
    \item{\code{fade_length}}{(Numeric scalar)
Chord fade-out time (seconds).}
    \item{\code{rise_length}}{(Numeric scalar)
Chord fade-in time (seconds).}
    \item{\code{num_harmonics}}{(Integerish scalar)
Number of harmonics (including the fundamental) to which
each tone should be expanded.}
    \item{\code{roll_off}}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}
    \item{\code{digits}}{Number of digits to which each partial's MIDI pitch should be rounded.}
  }}
}
\description{
Plays a chord as a wave file.
}
\seealso{
\code{\link{play_sox}}
}
