% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent.R
\name{represent}
\alias{represent}
\alias{represent.default}
\alias{represent.corpus}
\alias{represent.vec}
\alias{represent.coded_vec}
\alias{represent.list}
\title{Represent}
\usage{
represent(x, type, ...)

\method{represent}{default}(x, type, ...)

\method{represent}{corpus}(x, type, ...)

\method{represent}{vec}(x, type, ...)

\method{represent}{coded_vec}(x, type, ...)

\method{represent}{list}(x, type, ...)
}
\arguments{
\item{x}{Input object.}

\item{type}{Target type (see \code{\link{type}}).}

\item{...}{Further arguments to pass to \code{\link{represent}()}.}
}
\description{
Represents an object using a particular type.
}
