% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent.R
\name{representations}
\alias{representations}
\title{Representations}
\usage{
representations()
}
\value{
Character vector of representation labels.
}
\description{
Returns the available representations in the hrep package.
}
