% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{save_wav}
\alias{save_wav}
\title{Save wav file}
\usage{
save_wav(
  x,
  file,
  amplitude = 0.1,
  bit_depth = 16L,
  length_sec = 1,
  fade_length = 0.1,
  rise_length = 0.1,
  end_pad = 0.05,
  ...
)
}
\arguments{
\item{x}{Object to save; currently only individual chords are supported.
Chords are coerced to a \code{\link{wave}} representation.}

\item{file}{(Character scalar) Output file.}

\item{amplitude}{(Numeric scalar)
The wave is multiplied by this number before exporting.
The resulting wave should fall completely within the range [-1, 1].}

\item{bit_depth}{(Integer scalar)
The bit depth of the exported audio.}

\item{length_sec}{(Numeric scalar) Length of the output wave, in seconds.}

\item{fade_length}{(Numeric scalar)
Chord fade-out time (seconds).}

\item{rise_length}{(Numeric scalar)
Chord fade-in time (seconds).}

\item{end_pad}{(Numeric scalar)
Duration of silence (seconds) appended to the end of the audio file,
used to avoid clicks and other artifacts.}

\item{...}{
  Arguments passed on to \code{\link[=expand_harmonics]{expand_harmonics}}
  \describe{
    \item{\code{num_harmonics}}{(Integerish scalar)
Number of harmonics (including the fundamental) to which
each tone should be expanded.}
    \item{\code{roll_off}}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}
    \item{\code{digits}}{Number of digits to which each partial's MIDI pitch should be rounded.}
  }}
}
\description{
Saves object to a wav file by converting to the \code{\link{wave}} representation
and then writing to a wav file.
}
\seealso{
\code{\link{save_wav_sox}}
}
