% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-wav-sox.R
\name{save_wav_sox}
\alias{save_wav_sox}
\title{Save wav file (sox)}
\usage{
save_wav_sox(x, file, ...)
}
\arguments{
\item{x}{Object to save; methods exist for individual chords
and for vectors of chords (see \code{\link{vec}}).
Chords are coerced to a \code{\link{pi_chord}} representation.}

\item{file}{(Character scalar) Output file.}

\item{...}{Parameters passed to methods.}

\item{chord_length}{(Numeric scalar) Length of each output chord, in seconds.}
}
\description{
Saves object to a wav file using the 'sox' timbre from sox
(\url{http://sox.sourceforge.net/}).
}
\note{
The command-line sound-processing program sox
(\url{http://sox.sourceforge.net/})
must be installed and available on the command line
under the command \code{sox}.
}
