% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-spectrum.R
\name{set_labels}
\alias{set_labels}
\alias{set_labels.sparse_spectrum}
\title{Add labels}
\usage{
set_labels(x, labels)

\method{set_labels}{sparse_spectrum}(x, labels)
}
\arguments{
\item{x}{Object.}

\item{labels}{Character vector of labels to add to the object.
In the case of sparse spectra, there should be one label for each partial.}
}
\value{
The original object, with labels added.
}
\description{
Adds labels to an object, typically for plotting purposes.
}
\examples{
spectrum <- sparse_pi_spectrum("60 64 67", num_harmonics = 1)
labels <- as.character(c(1, 2, 3))
spectrum_with_labels <- set_labels(spectrum, labels)
plot(spectrum_with_labels)

}
