% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{silence}
\alias{silence}
\title{Silence}
\usage{
silence(duration, sample_rate = 44100L)
}
\arguments{
\item{duration}{(Numeric scalar)
Duration of the silence.}

\item{sample_rate}{(Integer scalar)
Sample rate.}
}
\description{
Creates a 'wave' object corresponding to silence.
}
