% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{size}
\alias{size}
\title{Size}
\usage{
size(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
The size of \code{x}, as an integer scalar.
}
\description{
Returns the size of an object, counted in symbols.
For an object of class \code{\link{vec}}, this will be the length of the vector.
For an object of class \code{\link{corpus}},
this will be the sum of the lengths of its constituent sequences.
}
