% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-pc-spectrum.R
\name{smooth_pc_spectrum}
\alias{smooth_pc_spectrum}
\alias{smooth_pc_spectrum.default}
\alias{smooth_pc_spectrum.sparse_pc_spectrum}
\title{Smooth pitch-class spectrum}
\usage{
smooth_pc_spectrum(x, sigma = 6.83, ...)

\method{smooth_pc_spectrum}{default}(x, sigma = 6.83,
  num_harmonics = 11L, roll_off = 1, ...)

\method{smooth_pc_spectrum}{sparse_pc_spectrum}(x, sigma = 6.83, ...)
}
\arguments{
\item{x}{Object to convert. By default \code{\link{sparse_pc_spectrum}}
is called first to convert the object to a sparse pitch-class spectrum.}

\item{sigma}{(Numeric scalar)
Standard deviation of the Gaussian distribution used to simulate
perceptual blurring. Defaults to 6.83 cents, after
\insertCite{Milne2016;textual}{hrep}.}

\item{...}{Provided for S3 method consistency.}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics (including the fundamental) to which
each tone should be expanded.}

\item{roll_off}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}
}
\description{
Creates an object of class \code{smooth_pc_spectrum},
describing a smooth pitch-class spectrum.
A smooth pitch-class spectrum describes perceptual weight
along a continuous pitch class scale.
The representation incorporates smoothing to account for
imprecisions in pitch perception.
}
\details{
This representation is created by expressing the sonority
as a sparse pitch-class spectrum (see \code{\link{sparse_pc_spectrum}})
and convolving the result with a Gaussian distribution with standard deviation
\code{sigma}.
}
\seealso{
This representation was inspired by \code{\link{milne_pc_spectrum}},
which embodies similar ideas with a slightly different implementation.
See \code{\link{smooth_pi_spectrum}} for an equivalent representation
in the pitch domain.
}
