% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-sparse-spectrum.R
\name{sparse_fr_spectrum}
\alias{sparse_fr_spectrum}
\alias{sparse_fr_spectrum.sparse_pi_spectrum}
\alias{sparse_fr_spectrum.pi_chord}
\alias{sparse_fr_spectrum.default}
\alias{sparse_fr_spectrum.list}
\alias{c.sparse_fr_spectrum}
\title{Sparse frequency spectrum}
\usage{
sparse_fr_spectrum(x, ...)

\method{sparse_fr_spectrum}{sparse_pi_spectrum}(x, ...)

\method{sparse_fr_spectrum}{pi_chord}(x, ...)

\method{sparse_fr_spectrum}{default}(x, ...)

\method{sparse_fr_spectrum}{list}(x, ...)

\method{c}{sparse_fr_spectrum}(..., x_digits = 6)
}
\arguments{
\item{x}{Input sonority.
\itemize{
\item Numeric vectors will be treated as vectors of MIDI note numbers,
and expanded into their implied harmonics.
\item Two-element lists will be treated as finalised spectra,
with the first element being a numeric vector of frequencies,
and the second element being a numeric vector of amplitudes.
}}

\item{...}{Further arguments passed to \code{\link{expand_harmonics}()},
depending on the method invoked.}

\item{x_digits}{(Integerish scalar) Number of significant digits
to which frequencies are rounded when being combined.}
}
\value{
An object of class \code{sparse_fr_spectrum}.
}
\description{
This function represents an input object as a
sparse frequency spectrum.
}
\details{
A sparse frequency spectrum comprises a finite set of spectral components,
each defined by a frequency (in Hz)
and an amplitude (expressed in arbitrary units, but with the
fundamental frequencies of chord pitches typically taking the value 1).

Sparse frequency spectra can be combined into one spectrum using \code{c(...)}.
Amplitudes are summed assuming incoherent wave superposition
(see \code{\link{sum_amplitudes}}).
}
