% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-pc-spectrum.R
\name{sparse_pc_spectrum}
\alias{sparse_pc_spectrum}
\alias{sparse_pc_spectrum.sparse_fr_spectrum}
\alias{sparse_pc_spectrum.list}
\alias{sparse_pc_spectrum.default}
\alias{sparse_pc_spectrum.pi_chord}
\title{Sparse pitch-class spectrum}
\usage{
sparse_pc_spectrum(x, ...)

\method{sparse_pc_spectrum}{sparse_fr_spectrum}(x, ...)

\method{sparse_pc_spectrum}{list}(x, ...)

\method{sparse_pc_spectrum}{default}(x, ...)

\method{sparse_pc_spectrum}{pi_chord}(x, amplitude = 1, ...)
}
\arguments{
\item{x}{Input sonority.}

\item{...}{Further arguments passed to \code{\link{expand_harmonics}()},
depending on the method invoked.}

\item{amplitude}{(Numeric vector)
Vector of amplitudes to assign to each pitch.
If a scalar value is provided, this value is assigned to all pitches}
}
\value{
An object of class \code{sparse_pc_spectrum}.
}
\description{
This function represents an input object as a sparse pitch-class spectrum.
}
\details{
A sparse pitch-class spectrum describes an input sonority as a finite set
of spectral components, each defined by a
pitch class (expressed as numbers in the range [0, 12))
and an amplitude (expressed in arbitrary units, but with the
fundamental frequencies of chord pitches typically taking the value 1).
}
