% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-pi-spectrum.R
\name{sparse_pi_spectrum}
\alias{sparse_pi_spectrum}
\alias{sparse_pi_spectrum.sparse_pi_spectrum}
\alias{sparse_pi_spectrum.sparse_fr_spectrum}
\alias{sparse_pi_spectrum.list}
\alias{sparse_pi_spectrum.default}
\alias{sparse_pi_spectrum.pi_chord}
\title{Sparse pitch spectrum}
\usage{
sparse_pi_spectrum(x, ...)

\method{sparse_pi_spectrum}{sparse_pi_spectrum}(x, ...)

\method{sparse_pi_spectrum}{sparse_fr_spectrum}(x, ...)

\method{sparse_pi_spectrum}{list}(x, ...)

\method{sparse_pi_spectrum}{default}(x, ...)

\method{sparse_pi_spectrum}{pi_chord}(x, amplitude = 1, ...)
}
\arguments{
\item{x}{Input sonority.}

\item{...}{Further arguments passed to \code{\link{expand_harmonics}()},
depending on the method invoked.}

\item{amplitude}{(Numeric vector)
Vector of amplitudes to assign to each pitch.
If a scalar value is provided, this value is assigned to all pitches}
}
\value{
An object of class \code{sparse_pi_spectrum}.
}
\description{
This function represents an input object as a sparse pitch spectrum.
}
\details{
A sparse pitch spectrum describes an input sonority as a finite set
of spectral components, each defined by a
pitch (expressed on the MIDI pitch scale)
and an amplitude (expressed in arbitrary units, but with the
fundamental frequencies of chord pitches typically taking the value 1).
}
