% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_spectrum_to_waveform}
\alias{sparse_spectrum_to_waveform}
\title{Note: this could be done more efficiently with ifft}
\usage{
sparse_spectrum_to_waveform(frequency, amplitude, seconds = 1,
  sample_rate = 44000, bit = 16)
}
\description{
Note: this could be done more efficiently with ifft
}
