% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-filter.R
\name{spectral_filter}
\alias{spectral_filter}
\title{Spectral filter}
\usage{
spectral_filter(x, fun, ...)
}
\arguments{
\item{x}{Object to filter, passed to \code{\link{sparse_fr_spectrum}}.}

\item{fun}{Function to apply. The input should be a vector of frequencies,
the output should be a vector of envelope amplitudes for those frequencies.
An amplitude of 1 means that the spectrum is left unchanged at that location.}

\item{...}{Optional parameters passed to \code{\link{sparse_fr_spectrum}}.}
}
\description{
Filters a sparse spectrum using a user-provided function.
}
