% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_amplitudes}
\alias{sum_amplitudes}
\title{Sum amplitudes}
\usage{
sum_amplitudes(x, y, coherent = FALSE)
}
\arguments{
\item{x}{First amplitude to sum (numeric vector)}

\item{y}{Second amplitude to sum (numeric vector)}

\item{coherent}{Whether or not the phases of the two tones are coherent (logical scalar)}
}
\description{
Sums amplitudes for pairs of pure tones (can be vectorised).
}
