% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_sound_levels}
\alias{sum_sound_levels}
\title{Sums pairs of sound levels assuming either coherent or incoherent (default) wave superposition, with dB input}
\usage{
sum_sound_levels(x, y, coherent = FALSE)
}
\arguments{
\item{x}{The first sound level to be summed in dB (can be vectorised)}

\item{y}{The second sound level to be summed in dB (can be vectorised)}
}
\description{
Sums pairs of sound levels assuming either coherent or incoherent (default) wave superposition, with dB input
}
