% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-sparse-spectrum.R
\name{to_dB}
\alias{to_dB}
\title{Given a frequency spectrum, re-express the amplitudes of each partial
in decibels.}
\usage{
to_dB(x, unit_amplitude_in_dB = 60, ...)
}
\arguments{
\item{x}{Frequency spectrum}

\item{unit_amplitude_in_dB}{An amplitude of 1 should be mapped
to this decibel value.}
}
\description{
Given a frequency spectrum, re-express the amplitudes of each partial
in decibels.
}
