% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R, R/generics.R, R/vec.R
\name{transform_symbols.coded_vec}
\alias{transform_symbols.coded_vec}
\alias{transform_symbols.corpus}
\alias{transform_symbols}
\alias{transform_symbols.vec}
\title{Transform symbols}
\usage{
\method{transform_symbols}{coded_vec}(x, f, type, ...)

\method{transform_symbols}{corpus}(x, f, type, progress = if
  (interactive()) "text" else "none", ...)

transform_symbols(x, f, type, ...)

\method{transform_symbols}{vec}(x, f, type, ...)
}
\arguments{
\item{x}{Object to transform.}

\item{f}{Function to apply to each symbol.}

\item{type}{Character scalar identifying the new symbol type.}

\item{...}{Further arguments to be passed to methods.}

\item{progress}{(Scalar character) Progress bar type, to be passed to
\code{\link[plyr]{llply}}.}
}
\value{
The transformed object.
}
\description{
Transforms the symbols within an object.
}
