% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transform_y}
\alias{transform_y}
\title{Transform y values}
\usage{
transform_y(x, f, y_unit, y_lab)
}
\arguments{
\item{x}{Object to transform.}

\item{f}{Function to apply (vectorised).}

\item{y_unit}{(Character scalar) New units for the y axis.}

\item{y_lab}{(Character scalar) New label for the y axis.}
}
\description{
Transforms an object's y values.
}
