% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{type}
\alias{type}
\title{Type}
\usage{
type(x)
}
\arguments{
\item{x}{Object whose type should be accessed.}
}
\value{
The type of \code{x}.
}
\description{
A \strong{type} is a way of representing a chord symbol.
}
\details{
Various types are available:
\itemize{
\item \code{\link{fr_chord}}
\item \code{\link{fr_sparse_spectrum}}
\item \code{\link{pi_chord}}
\item \code{\link{pi_sparse_spectrum}}
\item \code{\link{pc_chord}}
\item \code{\link{pc_chord_type}}
\item \code{\link{pc_set}}
\item \code{\link{pc_set_norm_form}}
\item \code{\link{pc_set_norm_order}}
\item \code{\link{pc_milne_spectrum}}
}

Various objects can be typed:
\itemize{
\item Individual chord objects, e.g. \code{\link{pc_set}}
\item Chord sequences, as created by \code{\link{vec}} and \code{\link{coded_vec}}
\item Corpora of chord sequences, as created by \code{\link{corpus}}
}
}
