% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{vec}
\alias{vec}
\alias{[.vec}
\alias{[<-.vec}
\alias{[[<-.vec}
\title{Vector}
\usage{
vec(x, type, metadata = list())

\method{[}{vec}(x, i)

\method{[}{vec}(x, i) <- value

\method{[[}{vec}(x, i) <- value
}
\arguments{
\item{x}{List of (un-encoded) chord symbols, each of class \code{type}.}

\item{type}{(Character scalar) Identifies the symbol \code{\link{type}}.}

\item{metadata}{(List) A (possibly-empty) list of metadata information.}

\item{i}{(Integerish vector) Index/indices to access.}

\item{value}{New value(s) to assign.}
}
\value{
An object of class "vec".
}
\description{
This function creates an object of the class "vec".
}
\details{
"vec" objects are typed vectors of chord symbols.
Their underlying representation is a list,
where every list element is of a specified type (see \code{\link{type}}).
}
\seealso{
\code{\link{coded_vec}}, code{\link{encode}}, code{\link{decode}}.
}
