% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pc-chord.R
\name{view}
\alias{view}
\alias{view.pc_chord}
\title{View}
\usage{
view(x, ...)

\method{view}{pc_chord}(x, ...)
}
\arguments{
\item{x}{Object to view.}

\item{...}{Parameters passed to methods.}
}
\description{
Views an object.
}
