% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{wave}
\alias{wave}
\alias{wave.default}
\alias{wave.sparse_fr_spectrum}
\title{Wave}
\usage{
wave(x, ...)

\method{wave}{default}(x, length_sec = 1, sample_rate = 44100, ...)

\method{wave}{sparse_fr_spectrum}(x, length_sec = 1,
  sample_rate = 44100, ...)
}
\arguments{
\item{x}{Input object.}

\item{...}{Arguments to be passed to \code{\link{sparse_fr_spectrum}}, as appropriate.}

\item{length_sec}{(Numeric scalar) Length of the output wave, in seconds.}

\item{sample_rate}{(Integerish scalar) The desired sample rate.}
}
\description{
This function represents an object as class "wave".
Under the hood, "wave" objects are numeric vectors
describing displacement as a function of time.
The sample rate can be accessed using the \code{\link{sample_rate}} accessor.
}
