% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R
\name{coded_vec}
\alias{coded_vec}
\title{Coded vector}
\usage{
coded_vec(x, type, metadata = list())
}
\arguments{
\item{x}{(Integer vector) The coded sequence.}

\item{type}{(Character scalar) Identifies the symbol \code{\link{type}}.}

\item{metadata}{(List) A (possibly-empty) list of metadata information.}
}
\value{
An object of class "coded_vec".
}
\description{
This function creates a coded vector.
Coded vectors are used to represent sequences of symbols
from a finite alphabet.
Each symbol should be coded as an integer using a bijective mapping.
}
\seealso{
\code{\link{vec}}.
}
