% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus}
\alias{corpus}
\alias{is.corpus}
\title{Corpus}
\usage{
corpus(x, type, metadata = list())

is.corpus(x, ...)
}
\arguments{
\item{x}{List of sequences, where each sequence is an object of class
\code{vec} or \code{coded_vec}.}

\item{type}{(Character scalar) Symbol type.}

\item{metadata}{(List) List of metadata information.}
}
\value{
An object of class \code{corpus}.
}
\description{
Creates \code{corpus} objects,
which comprise ordered sets of (possibly coded) vectors,
each describing a sequence of discrete symbols.
}
