% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R
\name{decode}
\alias{decode}
\alias{decode_pc_set}
\alias{decode_pc_chord_type}
\alias{decode_pc_chord}
\title{Decode}
\usage{
decode(x, x_type = type(x))

decode_pc_set(x)

decode_pc_chord_type(x)

decode_pc_chord(x)
}
\arguments{
\item{x}{Object to decode.}

\item{x_type}{(Character scalar)
The type that \code{x} is intended to represent (e.g. \code{"pc_chord"}).}
}
\value{
\code{decode} returns an object of class \code{vec};
the other functions return integer vectors.
}
\description{
Decodes an object from an integer-based encoding.
}
\details{
The \code{decode} function is vectorised and uses precomputed encodings.
The following functions represent alternative implementations
that are not vectorised and do not rely on precomputed encodings:
\itemize{
\item \code{decode_pc_set}
\item \code{decode_pc_chord_type}
\item \code{decode_pc_chord}
}
}
\seealso{
\code{\link{encode}} for the reverse operation.
}
