% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R
\name{decode}
\alias{decode}
\alias{decode.corpus}
\title{Decode}
\usage{
decode(x, x_type = type(x))

decode.corpus(x)
}
\arguments{
\item{x}{Object to decode.}
}
\description{
Decodes an object from an integer-based encoding.
}
\seealso{
\code{\link{encode}} for the reverse operation.
}
