% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-milne-spectrum.R
\name{.milne_pc_spectrum}
\alias{.milne_pc_spectrum}
\title{Constructor function for pitch-class Milne spectrum}
\usage{
.milne_pc_spectrum(x)
}
\arguments{
\item{x}{A numeric vector of pitch-class weights,
typically (but not necessarily) of length 1200.}
}
\value{
An object of class "milne_pc_spectrum".
}
\description{
This function constructs a "milne_pc_spectrum" object.
}
\seealso{
\code{\link{milne_pc_spectrum}}.
}
