% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R
\name{encode}
\alias{encode}
\alias{encode.vec}
\alias{encode.coded_vec}
\alias{encode.corpus}
\title{Encode}
\usage{
encode(x)

\method{encode}{vec}(x)

\method{encode}{coded_vec}(x)

\method{encode}{corpus}(x)
}
\arguments{
\item{x}{Object to transform.}
}
\value{
Encoded object.
}
\description{
Transforms a given object into an integer-based encoding.
}
\seealso{
\code{\link{decode}} for the reverse operation.
}
