% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-harmonics.R
\name{expand_harmonics}
\alias{expand_harmonics}
\title{Expand harmonics}
\usage{
expand_harmonics(x, ...)
}
\arguments{
\item{x}{Object whose harmonics should be expanded.}

\item{...}{Further parameters to be passed to
\code{\link{expand_harmonics.fr_sparse_spectrum}}.}
}
\description{
Expands each tone in an object into its implied harmonics.
}
