% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-harmonics.R
\name{expand_harmonics}
\alias{expand_harmonics}
\alias{expand_harmonics.sparse_fr_spectrum}
\alias{expand_harmonics.sparse_pi_spectrum}
\alias{expand_harmonics.pi_chord}
\title{Expand harmonics}
\usage{
expand_harmonics(x, num_harmonics = 11L, roll_off = 1, digits = 6)

\method{expand_harmonics}{sparse_fr_spectrum}(x, num_harmonics = 11L,
  roll_off = 1, digits = 6)

\method{expand_harmonics}{sparse_pi_spectrum}(x, num_harmonics = 11L,
  roll_off = 1, digits = 6)

\method{expand_harmonics}{pi_chord}(x, num_harmonics = 11L,
  roll_off = 1, digits = 6)
}
\arguments{
\item{x}{Object whose harmonics should be expanded.
Should be of class
\code{sparse_pi_spectrum},
\code{sparse_fr_spectrum}, or
\code{pi_chord}.}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics (including the fundamental) to which
each tone should be expanded.}

\item{roll_off}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}

\item{digits}{Number of digits to which each partial's MIDI pitch should be rounded.}
}
\description{
Expands each tone in an object into its implied harmonics.
}
