% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-filter.R
\name{filter_spectrum}
\alias{filter_spectrum}
\title{Filter spectrum}
\usage{
filter_spectrum(x, fun, ...)
}
\arguments{
\item{x}{Object to filter, passed to \code{\link{sparse_fr_spectrum}}.}

\item{fun}{Function to apply. The input should be a vector of frequencies,
the output should be a vector of envelope amplitudes for those frequencies.
The spectral amplitudes will be multiplied by the outputs of this function
for each frequency.}

\item{...}{Optional parameters passed to \code{\link{sparse_fr_spectrum}}.}
}
\description{
Filters a sparse spectrum using a user-provided function.
}
