% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-spectrum.R
\name{get_complex_tone}
\alias{get_complex_tone}
\title{Get complex tone}
\usage{
get_complex_tone(fundamental_pc, weight, array_dim, num_harmonics, rho,
  sigma)
}
\description{
Wrapper for \code{new_complex_tone} that implements caching.
}
