% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_harmonic_template}
\alias{get_harmonic_template}
\title{Get harmonic template}
\usage{
get_harmonic_template(num_harmonics, amplitude,
  roll_off = get_midi_params()$roll_off, interval_scale = "ratio",
  round_midi_intervals = TRUE)
}
\arguments{
\item{num_harmonics}{Number of harmonics (including fundamental)}

\item{amplitude}{amplitude of the fundamental frequency}

\item{roll_off}{Roll-off constant}

\item{interval_scale}{Can be "ratio" or "midi"}
}
\value{
\code{data.frame} of frequencies and amplitudes
}
\description{
Gets a harmonic template, using non-stretched octaves.
}
