% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrep-package.R
\docType{package}
\name{hrep-package}
\alias{hrep}
\alias{hrep-package}
\title{hrep: Harmony Representations}
\description{
This package provides utilities for representing and manipulating 
    chord sequences for perceptually informed harmony modelling.
    The available representations include symbolic representations,
    acoustic representations, and sensory representations.
    Integer encodings are defined for several symbolic representations,
    allowing the user to efficiently express chord sequences as sequences
    of integers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmcharrison/hrep}
  \item Report bugs at \url{https://github.com/pmcharrison/hrep/issues}
}

}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
