% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrep-package.R
\docType{package}
\name{hrep-package}
\alias{hrep}
\alias{hrep-package}
\title{hrep: Harmony Representations}
\description{
This package provides methods for representing and manipulating chord sequences.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
