% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-chord.R
\name{is.pc_chord}
\alias{is.pc_chord}
\title{Pitch-class chord type checking}
\usage{
is.pc_chord(x)

is.pc_chord(x)
}
\arguments{
\item{x}{Object to check.}

\item{x}{Object to analyse}
}
\value{
Scalar logical.

Logical scalar.
}
\description{
Checks whether an object is of class "pc_chord".

Checks whether an object is of class "pc_chord".
}
