% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded-vec.R, R/corpus.R, R/generics.R, R/vec.R
\name{metadata.coded_vec}
\alias{metadata.coded_vec}
\alias{metadata.corpus}
\alias{metadata<-.corpus}
\alias{metadata}
\alias{metadata.default}
\alias{metadata<-}
\alias{metadata.vec}
\title{Metadata}
\usage{
\method{metadata}{coded_vec}(x)

\method{metadata}{corpus}(x)

\method{metadata}{corpus}(x) <- value

metadata(x)

\method{metadata}{default}(x)

metadata(x) <- value

\method{metadata}{vec}(x)
}
\arguments{
\item{x}{Method whose metadata should be accessed.}
}
\description{
Additional information about an object may be stored in its
metadata attribute.
}
