% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set.R
\name{pc_set}
\alias{pc_set}
\alias{pc_set.numeric}
\alias{pc_set.pc_set}
\alias{pc_set.pc_chord}
\alias{pc_set.pi_chord}
\alias{pc_set.fr_chord}
\alias{pc_set.pc_set_norm_order}
\alias{pc_set.pc_set_type}
\alias{pc_set.character}
\title{Pitch-class set}
\usage{
pc_set(x)

\method{pc_set}{numeric}(x)

\method{pc_set}{pc_set}(x)

\method{pc_set}{pc_chord}(x)

\method{pc_set}{pi_chord}(x)

\method{pc_set}{fr_chord}(x)

\method{pc_set}{pc_set_norm_order}(x)

\method{pc_set}{pc_set_type}(x)

\method{pc_set}{character}(x)
}
\arguments{
\item{x}{Object to represent as a pitch-class set.}
}
\value{
Returns an object of class \code{pc_set}.
}
\description{
This function represents an object as a pitch-class set.
}
