% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-spectrum.R
\name{pc_spectrum_template_1}
\alias{pc_spectrum_template_1}
\title{Pitch-class spectrum, template 1
Makes a Gaussian pitch-class  spectral template with unit mass, centred on 0,
with standard deviation <sigma>.
The template will be truncated to zero for points <truncation-point>
standard deviations or further away from the mean,
after \insertCite{Milne2016a;textual}{hrep}.}
\usage{
pc_spectrum_template_1(array_dim, sigma, truncation_point)
}
\description{
Pitch-class spectrum, template 1
Makes a Gaussian pitch-class  spectral template with unit mass, centred on 0,
with standard deviation <sigma>.
The template will be truncated to zero for points <truncation-point>
standard deviations or further away from the mean,
after \insertCite{Milne2016a;textual}{hrep}.
}
\references{
\insertAllCited{}
}
