% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{sum_amplitudes}
\alias{sum_amplitudes}
\title{Sum amplitudes}
\usage{
sum_amplitudes(x, y, coherent = FALSE, dB = FALSE)
}
\arguments{
\item{x}{(Numeric vector) First amplitude to sum.}

\item{y}{(Numeric vector) Second amplitude to sum.}

\item{coherent}{(Logical scalar) Whether or not the phases of the two tones are coherent.}

\item{dB}{(Logical scalar) Whether or not the amplitudes are provided in decibels.}
}
\description{
Sums amplitudes for pairs of pure tones (vectorised).
}
