% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pc-chord.R, R/pc-set.R,
%   R/pi-chord.R, R/vec.R
\name{view}
\alias{view}
\alias{view.pc_chord}
\alias{view.pc_set}
\alias{view.pi_chord}
\alias{view.vec}
\title{View}
\usage{
view(x, ...)

\method{view}{pc_chord}(x, ...)

\method{view}{pc_set}(x, ...)

\method{view}{pi_chord}(x, ...)

\method{view}{vec}(x, annotate = NULL, ...)
}
\arguments{
\item{x}{Object to view.}

\item{...}{Parameters passed to methods.}

\item{annotate}{An optional character vector of the same length
as \code{x}, with which to annotate each element.}
}
\description{
Views an object.
}
