% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{wave}
\alias{wave}
\alias{wave.default}
\alias{wave.sparse_fr_spectrum}
\title{Wave}
\usage{
wave(x, ...)

\method{wave}{default}(
  x,
  length_sec = 1,
  sample_rate = 44100,
  rise_length = 0,
  fade_length = 0,
  ...
)

\method{wave}{sparse_fr_spectrum}(
  x,
  length_sec = 1,
  sample_rate = 44100,
  rise_length = 0,
  fade_length = 0,
  phase = 0,
  ...
)
}
\arguments{
\item{x}{Input object.}

\item{...}{
  Arguments passed on to \code{\link[=expand_harmonics]{expand_harmonics}}
  \describe{
    \item{\code{num_harmonics}}{(Integerish scalar)
Number of harmonics (including the fundamental) to which
each tone should be expanded.}
    \item{\code{roll_off}}{(Numeric scalar) Parametrises the amount of amplitude roll-off
in the harmonics, with greater values corresponding to higher roll-off.}
    \item{\code{digits}}{Number of digits to which each partial's MIDI pitch should be rounded.}
    \item{\code{label_harmonics}}{If TRUE, then the harmonics in the resulting spectrum are labelled with their harmonic numbers.}
  }}

\item{length_sec}{(Numeric scalar) Length of the output wave, in seconds.}

\item{sample_rate}{(Integerish scalar) The desired sample rate.}

\item{rise_length}{(Numeric scalar)
Chord fade-in time (seconds).}

\item{fade_length}{(Numeric scalar)
Chord fade-out time (seconds).}

\item{phase}{(Numeric scalar)
Single number specifying the phase of the underlying sine waves to be used.}
}
\description{
This function represents an object as class "wave".
Under the hood, "wave" objects are numeric vectors
describing displacement as a function of time.
The sample rate can be accessed using the \code{\link{sample_rate}} accessor.
}
